+++ Tests/SuperwallKitTests/Analytics/Internal Tracking/TrackTests.swift
@@ -288,12 +288,18 @@ final class TrackingTests: XCTestCase {
   func test_triggerFire_noRuleMatch() async {
     let triggerName = "My Trigger"
     let dependencyContainer = DependencyContainer()
-    let result = await Superwall.shared.track(InternalSuperwallEvent.TriggerFire(triggerResult: .noRuleMatch, triggerName: triggerName, sessionEventsManager: dependencyContainer.sessionEventsManager))
+    let unmatchedRules: [UnmatchedRule] = [
+      .init(source: .expression, experimentId: "1"),
+      .init(source: .occurrence, experimentId: "2")
+    ]
+    let result = await Superwall.shared.track(InternalSuperwallEvent.TriggerFire(triggerResult: .noRuleMatch(unmatchedRules), triggerName: triggerName, sessionEventsManager: dependencyContainer.sessionEventsManager))
     XCTAssertNotNil(result.parameters.eventParams["$app_session_id"])
     XCTAssertTrue(result.parameters.eventParams["$is_standard_event"] as! Bool)
     XCTAssertEqual(result.parameters.eventParams["$event_name"] as! String, "trigger_fire")
     XCTAssertEqual(result.parameters.eventParams["$result"] as! String, "no_rule_match")
     XCTAssertEqual(result.parameters.eventParams["$trigger_name"] as! String, triggerName)
+    XCTAssertEqual(result.parameters.eventParams["$unmatched_rule_1"] as! String, "EXPRESSION")
+    XCTAssertEqual(result.parameters.eventParams["$unmatched_rule_2"] as! String, "OCCURRENCE")
     // TODO: Missing test for trigger_session_id here. Need to figure out a way to activate it
   }